feat!: add more unknown changes

+++ routes/anime.py
@@ -1,14 +1,20 @@
-from fastapi import APIRouter
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException, status
 
 from database import postgres
 from config import API_PREFIX
-from schemas import AddAnimePayload, GenreID, AnimeID
+from schemas import AddAnimePayload, GenreID, AnimeID, UserID
+import dependencies
 
 router = APIRouter(prefix=API_PREFIX)
 
 
 @router.post("/animes")
-async def add_anime(anime: AddAnimePayload) -> AnimeID:
+async def add_anime(anime: AddAnimePayload, user: Annotated[UserID, Depends(dependencies.get_current_user)]) -> AnimeID:
+    if not user.can_manage_anime():
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions to do this action")
+
     anime_id = await postgres.add_anime(
         anime.name,
         anime.mal_id,
@@ -26,12 +32,16 @@ async def get_anime(anime_id: int) -> AnimeID:
 
 
 @router.delete("/animes/{anime_id}")
-async def delete_anime(anime_id: int):
+async def delete_anime(anime_id: int, user: Annotated[UserID, Depends(dependencies.get_current_user)]):
+    if not user.can_manage_anime():
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions to do this action")
     await postgres.delete_anime(anime_id)
 
 
 @router.post("/animes/genres")
-async def add_genre(genre_name: str) -> GenreID:
+async def add_genre(genre_name: str, user: Annotated[UserID, Depends(dependencies.get_current_user)]) -> GenreID:
+    if not user.can_manage_anime():
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions to do this action")
     genre_id = await postgres.add_genre(genre_name)
     return GenreID(
         name=genre_name,
